fix: campaigns now shows loading correctly
